refactor: inline ScreenshotNamerDSL into DSL module
8-line module in a separate file with pure delegation.
Move screenshot_section, screenshot_group, and screenshot_namer
directly into DSL. Delete the file.
